refactor: extract ImageLightbox to shared component
Both attachment-list.tsx and attachment-context.tsx had duplicate
ImageLightbox implementations. Extracted to a single shared component.
